# Remove unnecessary copy while decoding and constructing string