--- a/pkg/ottl/ottlfuncs/func_decode.go
+++ b/pkg/ottl/ottlfuncs/func_decode.go
@@ -7,11 +7,8 @@ import (
 	"context"
 	"encoding/base64"
 	"fmt"
-	"strings"
 
 	"go.opentelemetry.io/collector/pdata/pcommon"
-	"golang.org/x/text/encoding"
-	"golang.org/x/text/encoding/ianaindex"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal/textutils"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/ottl"
@@ -71,7 +68,7 @@ func Decode[K any](target ottl.Getter[K], encoding string) (ottl.ExprFunc[K], er
 			}
 			return string(decodedBytes), nil
 		default:
-			e, err := getEncoding(encoding)
+			e, err := textutils.LookupEncoding(encoding)
 			if err != nil {
 				return nil, err
 			}
@@ -84,20 +81,4 @@ func Decode[K any](target ottl.Getter[K], encoding string) (ottl.ExprFunc[K], er
 			return decodedString, nil
 		}
 	}, nil
-}
-
-func getEncoding(encoding string) (encoding.Encoding, error) {
-	if e, ok := textutils.EncodingOverridesMap.Get(strings.ToLower(encoding)); ok {
-		return e, nil
-	}
-	e, err := ianaindex.IANA.Encoding(encoding)
-	if err != nil {
-		return nil, fmt.Errorf("could not get encoding for %s: %w", encoding, err)
-	}
-	if e == nil {
-		// for some encodings a nil error and a nil encoding is returned, so we need to double check
-		// if the encoding is actually set here
-		return nil, fmt.Errorf("no decoder available for encoding: %s", encoding)
-	}
-	return e, nil
 }