--- a/pkg/stanza/fileconsumer/internal/header/reader.go
+++ b/pkg/stanza/fileconsumer/internal/header/reader.go
@@ -44,15 +44,15 @@ func NewReader(set component.TelemetrySettings, cfg Config) (*Reader, error) {
 
 // Process checks if the given token is a line of the header, and consumes it if it is.
 // An EndOfHeaderError is returned if the given line was not a header line.
-func (r *Reader) Process(ctx context.Context, token []byte, fileAttributes map[string]any) error {
-	if !r.cfg.regex.Match(token) {
+func (r *Reader) Process(ctx context.Context, token string, fileAttributes map[string]any) error {
+	if !r.cfg.regex.MatchString(token) {
 		return ErrEndOfHeader
 	}
 
 	firstOperator := r.pipeline.Operators()[0]
 
 	newEntry := entry.New()
-	newEntry.Body = string(token)
+	newEntry.Body = token
 
 	if err := firstOperator.Process(ctx, newEntry); err != nil {
 		r.set.Logger.Error("process header entry", zap.Error(err))