--- a/pkg/stanza/operator/input/tcp/input.go
+++ b/pkg/stanza/operator/input/tcp/input.go
@@ -20,7 +20,7 @@ import (
 	"go.uber.org/zap"
 	"golang.org/x/text/encoding"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/decode"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal/textutils"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/helper"
 )
@@ -129,7 +129,7 @@ func (i *Input) goHandleMessages(ctx context.Context, conn net.Conn, cancel cont
 		defer i.wg.Done()
 		defer cancel()
 
-		dec := decode.New(i.encoding)
+		dec := i.encoding.NewDecoder()
 		if i.OneLogPerPacket {
 			var buf bytes.Buffer
 			_, err := io.Copy(&buf, conn)
@@ -158,14 +158,14 @@ func (i *Input) goHandleMessages(ctx context.Context, conn net.Conn, cancel cont
 	}()
 }
 
-func (i *Input) handleMessage(ctx context.Context, conn net.Conn, dec *decode.Decoder, log []byte) {
-	decoded, err := dec.Decode(log)
+func (i *Input) handleMessage(ctx context.Context, conn net.Conn, dec *encoding.Decoder, log []byte) {
+	decoded, err := textutils.DecodeAsString(dec, log)
 	if err != nil {
 		i.Logger().Error("Failed to decode data", zap.Error(err))
 		return
 	}
 
-	entry, err := i.NewEntry(string(decoded))
+	entry, err := i.NewEntry(decoded)
 	if err != nil {
 		i.Logger().Error("Failed to create entry", zap.Error(err))
 		return