--- a/pkg/stanza/operator/input/udp/input.go
+++ b/pkg/stanza/operator/input/udp/input.go
@@ -15,7 +15,7 @@ import (
 	"go.uber.org/zap"
 	"golang.org/x/text/encoding"
 
-	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/decode"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal/textutils"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/operator/helper"
 )
@@ -86,7 +86,7 @@ func (i *Input) goHandleMessages(ctx context.Context) {
 func (i *Input) readAndProcessMessages(ctx context.Context) {
 	defer i.wg.Done()
 
-	dec := decode.New(i.encoding)
+	dec := i.encoding.NewDecoder()
 	readBuffer := make([]byte, MaxUDPSize)
 	scannerBuffer := make([]byte, 0, MaxUDPSize)
 	for {
@@ -107,7 +107,7 @@ func (i *Input) readAndProcessMessages(ctx context.Context) {
 	}
 }
 
-func (i *Input) processMessage(ctx context.Context, message []byte, remoteAddr net.Addr, dec *decode.Decoder, scannerBuffer []byte) {
+func (i *Input) processMessage(ctx context.Context, message []byte, remoteAddr net.Addr, dec *encoding.Decoder, scannerBuffer []byte) {
 	if i.OneLogPerPacket {
 		log := truncateMaxLog(message)
 		i.handleMessage(ctx, remoteAddr, dec, log)
@@ -158,7 +158,7 @@ func (i *Input) readMessagesAsync(ctx context.Context) {
 func (i *Input) processMessagesAsync(ctx context.Context) {
 	defer i.wg.Done()
 
-	dec := decode.New(i.encoding)
+	dec := i.encoding.NewDecoder()
 	scannerBuffer := make([]byte, 0, MaxUDPSize)
 
 	for {
@@ -186,18 +186,14 @@ func truncateMaxLog(data []byte) (token []byte) {
 	return data
 }
 
-func (i *Input) handleMessage(ctx context.Context, remoteAddr net.Addr, dec *decode.Decoder, log []byte) {
-	decoded := log
-	if i.encoding != encoding.Nop {
-		var err error
-		decoded, err = dec.Decode(log)
-		if err != nil {
-			i.Logger().Error("Failed to decode data", zap.Error(err))
-			return
-		}
+func (i *Input) handleMessage(ctx context.Context, remoteAddr net.Addr, dec *encoding.Decoder, log []byte) {
+	decoded, err := textutils.DecodeAsString(dec, log)
+	if err != nil {
+		i.Logger().Error("Failed to decode data", zap.Error(err))
+		return
 	}
 
-	entry, err := i.NewEntry(string(decoded))
+	entry, err := i.NewEntry(decoded)
 	if err != nil {
 		i.Logger().Error("Failed to create entry", zap.Error(err))
 		return