--- a/receiver/kafkareceiver/kafka_receiver_test.go
+++ b/receiver/kafkareceiver/kafka_receiver_test.go
@@ -1037,11 +1037,9 @@ func TestLogsConsumerGroupHandler_unmarshal_text(t *testing.T) {
 				assert.NoError(t, err)
 				wg.Done()
 			}()
-			encCfg := textutils.NewEncodingConfig()
-			encCfg.Encoding = test.enc
-			enc, err := encCfg.Build()
+			enc, err := textutils.LookupEncoding(test.enc)
 			require.NoError(t, err)
-			encoder := enc.Encoding.NewEncoder()
+			encoder := enc.NewEncoder()
 			encoded, err := encoder.Bytes([]byte(test.text))
 			require.NoError(t, err)
 			t1 := time.Now()