--- a/receiver/kafkareceiver/text_unmarshaler.go
+++ b/receiver/kafkareceiver/text_unmarshaler.go
@@ -8,12 +8,13 @@ import (
 
 	"go.opentelemetry.io/collector/pdata/pcommon"
 	"go.opentelemetry.io/collector/pdata/plog"
+	"golang.org/x/text/encoding"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal/textutils"
 )
 
 type textLogsUnmarshaler struct {
-	enc *textutils.Encoding
+	decoder *encoding.Decoder
 }
 
 func newTextLogsUnmarshaler() LogsUnmarshalerWithEnc {
@@ -21,18 +22,18 @@ func newTextLogsUnmarshaler() LogsUnmarshalerWithEnc {
 }
 
 func (r *textLogsUnmarshaler) Unmarshal(buf []byte) (plog.Logs, error) {
-	if r.enc == nil {
+	if r.decoder == nil {
 		return plog.Logs{}, errors.New("encoding not set")
 	}
 	p := plog.NewLogs()
-	decoded, err := r.enc.Decode(buf)
+	decoded, err := textutils.DecodeAsString(r.decoder, buf)
 	if err != nil {
 		return p, err
 	}
 
 	l := p.ResourceLogs().AppendEmpty().ScopeLogs().AppendEmpty().LogRecords().AppendEmpty()
 	l.SetObservedTimestamp(pcommon.NewTimestampFromTime(time.Now()))
-	l.Body().SetStr(string(decoded))
+	l.Body().SetStr(decoded)
 	return p, nil
 }
 
@@ -42,13 +43,11 @@ func (r *textLogsUnmarshaler) Encoding() string {
 
 func (r *textLogsUnmarshaler) WithEnc(encodingName string) (LogsUnmarshalerWithEnc, error) {
 	var err error
-	encCfg := textutils.NewEncodingConfig()
-	encCfg.Encoding = encodingName
-	enc, err := encCfg.Build()
+	enc, err := textutils.LookupEncoding(encodingName)
 	if err != nil {
 		return nil, err
 	}
 	return &textLogsUnmarshaler{
-		enc: &enc,
+		decoder: enc.NewDecoder(),
 	}, nil
 }