… .chloggen/rm-unncessary-copy-enhance.yaml
… .chloggen/rm-unncessary-copy.yaml
… pkg/stanza/fileconsumer/internal/header/reader_test.go
… pkg/stanza/fileconsumer/internal/reader/reader_test.go