Before (including copy of the bytes needed upstream):
```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/fileconsumer/internal/reader
cpu: Apple M2 Max
BenchmarkFileRead
BenchmarkFileRead-12    	    6890	    159408 ns/op	  125802 B/op	     132 allocs/op
PASS
```

After
```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/fileconsumer/internal/reader
cpu: Apple M2 Max
BenchmarkFileRead
BenchmarkFileRead-12    	    7662	    154923 ns/op	  121656 B/op	     130 allocs/op
PASS
```