@atoulme 2025-02-06
Changes look good, do you know any benchmarks that might be affected worth running to see the impact?