@andrzej-stencel 2025-02-10
@bogdandrutu I'm not sure I understand the benchmark results you posted. When I ran the `BenchmarkFileRead` you created before and after the changes on this PR, I'm getting different results. Am I missing something?

Here's what I did:

```console
$ git checkout 0820ea0d51e2d9acd2649911bdb06ae16a558f12 # commit on main just before this change
$ # copy the BenchmarkFileRead code into `reader_test.go`
$ (cd pkg/stanza/fileconsumer/internal/reader && go test -benchmem -run=^$ -bench ^BenchmarkFileRead$ . -v -count=10 | tee /tmp/fileread-$(git rev-parse --short HEAD).txt)
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/fileconsumer/internal/reader
BenchmarkFileRead
BenchmarkFileRead-12                9499            125685 ns/op           23288 B/op         31 allocs/op
BenchmarkFileRead-12                8906            126189 ns/op           23288 B/op         31 allocs/op
BenchmarkFileRead-12                9572            127992 ns/op           23288 B/op         31 allocs/op
BenchmarkFileRead-12                8439            126469 ns/op           23288 B/op         31 allocs/op
BenchmarkFileRead-12                9093            126672 ns/op           23288 B/op         31 allocs/op
BenchmarkFileRead-12                9602            125069 ns/op           23288 B/op         31 allocs/op
BenchmarkFileRead-12                9024            125003 ns/op           23288 B/op         31 allocs/op
BenchmarkFileRead-12                9194            126363 ns/op           23288 B/op         31 allocs/op
BenchmarkFileRead-12                8574            126052 ns/op           23288 B/op         31 allocs/op
BenchmarkFileRead-12                9513            127504 ns/op           23288 B/op         31 allocs/op
PASS
ok      github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/fileconsumer/internal/reader       12.965s
$ git reset --hard # remove the changed code before checking out next commit
$ git checkout 4244166551b2b42c824de934009a7ce535b0e656 # commit on main with changes from this PR
$ (cd pkg/stanza/fileconsumer/internal/reader && go test -benchmem -run=^$ -bench ^BenchmarkFileRead$ . -v -count=10 | tee /tmp/fileread-$(git rev-parse --short HEAD).txt)
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/fileconsumer/internal/reader
BenchmarkFileRead
BenchmarkFileRead-12                8442            138627 ns/op          121657 B/op        130 allocs/op
BenchmarkFileRead-12                8010            138947 ns/op          121656 B/op        130 allocs/op
BenchmarkFileRead-12                8406            139942 ns/op          121656 B/op        130 allocs/op
BenchmarkFileRead-12                8325            139564 ns/op          121656 B/op        130 allocs/op
BenchmarkFileRead-12                8341            139051 ns/op          121656 B/op        130 allocs/op
BenchmarkFileRead-12                8055            140335 ns/op          121656 B/op        130 allocs/op
BenchmarkFileRead-12                8577            143245 ns/op          121656 B/op        130 allocs/op
BenchmarkFileRead-12                8512            139856 ns/op          121656 B/op        130 allocs/op
BenchmarkFileRead-12                8336            138491 ns/op          121656 B/op        130 allocs/op
BenchmarkFileRead-12                8548            139459 ns/op          121656 B/op        130 allocs/op
PASS
ok      github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/fileconsumer/internal/reader       13.213s
``` 

These benchmarks differ in allocations:

```console
$ benchstat /tmp/fileread-0820ea0d51.txt /tmp/fileread-4244166551.txt
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-collector-contrib/pkg/stanza/fileconsumer/internal/reader
            │ /tmp/fileread-0820ea0d51.txt │    /tmp/fileread-4244166551.txt     │
            │            sec/op            │   sec/op     vs base                │
FileRead-12                    126.3µ ± 1%   139.5µ ± 1%  +10.48% (p=0.000 n=10)

            │ /tmp/fileread-0820ea0d51.txt │      /tmp/fileread-4244166551.txt      │
            │             B/op             │     B/op       vs base                 │
FileRead-12                   22.74Ki ± 0%   118.80Ki ± 0%  +422.40% (p=0.000 n=10)

            │ /tmp/fileread-0820ea0d51.txt │     /tmp/fileread-4244166551.txt     │
            │          allocs/op           │  allocs/op   vs base                 │
FileRead-12                     31.00 ± 0%   130.00 ± 0%  +319.35% (p=0.000 n=10)
```