@djaglowski 2025-02-10
```
            │ /tmp/before.txt │           /tmp/after.txt            │
            │     sec/op      │   sec/op     vs base                │
FileRead-14       636.3µ ± 0%   633.6µ ± 1%  -0.42% (p=0.041 n=100)

            │ /tmp/before.txt │             /tmp/after.txt              │
            │      B/op       │     B/op       vs base                  │
FileRead-14      22.77Ki ± 0%   118.84Ki ± 0%  +421.82% (p=0.000 n=100)

            │ /tmp/before.txt │            /tmp/after.txt             │
            │    allocs/op    │  allocs/op   vs base                  │
FileRead-14        30.00 ± 0%   130.00 ± 0%  +333.33% (p=0.000 n=100)
```

I see nearly identical memory and allocation increases as @andrzej-stencel.

I think we should revert the change unless there is an easy fix. Automatic reverting is not possible because of subsequent changes. @bogdandrutu